restructure and other changes to pack this into a pip installable
