--- CLUE_PYTHON.egg-info/SOURCES.txt
@@ -1,20 +1,0 @@
-LICENSE
-README.md
-pyproject.toml
-setup.py
-CLUE_PYTHON.egg-info/PKG-INFO
-CLUE_PYTHON.egg-info/SOURCES.txt
-CLUE_PYTHON.egg-info/dependency_links.txt
-CLUE_PYTHON.egg-info/requires.txt
-CLUE_PYTHON.egg-info/top_level.txt
-src/__init__.py
-src/concepts.py
-src/config.py
-src/datatypes.py
-src/evaluate.py
-src/example_context_usability.py
-src/example_model_uncertainty.py
-src/models.py
-src/output_sequence_generator.py
-src/uncertainty.py
-tests/__init__.py
