+++ src/llm_uncertainty_evals/datatypes.py
